+++ Changes.md
@@ -8,6 +8,7 @@ Version 3.3.1.0
   5. One section only.
   6. Remove 'coursedisplay' setting favour of combining the two and adding a link on the right of the toggle.
   7. Add opactity option for all four toggle colour settings.
+  8. private or protected - #31.
 
 Version 3.3.0.2
   1. Make the 'readme' course setting only display and not store in the database, thus regenerate and not broken in the